Moving emails to top level
