+++ src/Email.tsx
@@ -3,9 +3,9 @@ import { renderToStaticMarkup } from "react-dom/server";
 import { Fronts, Front } from "./api";
 import { Banner } from "./components/Banner";
 import { Collections } from "./components/Collections";
-import { FilmToday } from "./components/emails/film-today/FilmToday";
-import { MediaBriefing } from "./components/emails/media-briefing/MediaBriefing";
-import { Opinion } from "./components/emails/opinion/Opinion";
+import { FilmToday } from "./emails/film-today/FilmToday";
+import { MediaBriefing } from "./emails/media-briefing/MediaBriefing";
+import { Opinion } from "./emails/opinion/Opinion";
 import { Footer } from "./components/Footer";
 import { Center } from "./layout/Center";
 import { default as minifyCssString } from "minify-css-string";
