+++ src/components/buttons/ContinueButton.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { LinkCSS, ImageCSS } from "../css";
-import { TableRowCell } from "../layout/Table";
+import { LinkCSS, ImageCSS } from "../../css";
+import { TableRowCell } from "../../layout/Table";
 
 interface Props {
     label: string;
