+++ src/emails/opinion/VariantB.tsx
@@ -1,16 +1,16 @@
 import React from "react";
-import { Collection as ICollection } from "../../../api";
-import { TableRowCell } from "../../../layout/Table";
+import { Collection as ICollection } from "../../api";
+import { TableRowCell } from "../../layout/Table";
 import {
     CommentCollection,
     EditorialCollection,
     MediaCollection,
     LinkCollection,
     DefaultCollection
-} from "./collections/commentC/Collection";
-import { getDesignType } from "../../../utils/getDesignType";
+} from "./collections/opinionB/Collection";
+import { getDesignType } from "../../utils/getDesignType";
 
-export const VariantC: React.FC<{
+export const VariantB: React.FC<{
     frontId: string;
     collections: ICollection[];
     salt: string;
@@ -46,7 +46,6 @@ export const VariantC: React.FC<{
                 );
             case "media":
                 return <MediaCollection collection={collection} salt={salt} />;
-
             case "link":
                 return <LinkCollection collection={collection} salt={salt} />;
             case "default":
