+++ src/emails/opinion/collections/opinionB/MediaCard.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { FontCSS, TdCSS, TableCSS, ImageCSS } from "../../../../../css";
+import { FontCSS, TdCSS, TableCSS, ImageCSS } from "../../../../css";
 import { palette } from "@guardian/src-foundations";
-import { Content } from "../../../../../api";
-import { formatImage } from "../../../../../image";
-import { Table, RowCell, TableRowCell } from "../../../../../layout/Table";
-import { Padding } from "../../../../../layout/Padding";
+import { Content } from "../../../../api";
+import { formatImage } from "../../../../image";
+import { Table, RowCell, TableRowCell } from "../../../../layout/Table";
+import { Padding } from "../../../../layout/Padding";
 
 const tdStyle: TdCSS = {
     backgroundColor: palette.neutral[20],
